done with sem7

+++ ass7/matrix-matrix.cpp
@@ -11,9 +11,9 @@ int main(int argc, char **argv)
   MPI_Comm_size(MPI_COMM_WORLD, &size);
 
   int N = 500;
-  std::vector<int> A(N * N);
-  std::vector<int> B(N * N);
-  std::vector<int> C(N * N, 0);
+  std::vector<long long> A(N * N);
+  std::vector<long long> B(N * N);
+  std::vector<long long> C(N * N, 0);
 
   if (rank == 0)
   {
@@ -23,14 +23,14 @@ int main(int argc, char **argv)
       B[i] = i + 1;
   }
 
-  MPI_Bcast(B.data(), N * N, MPI_INT, 0, MPI_COMM_WORLD);
+  MPI_Bcast(B.data(), N * N, MPI_LONG_LONG, 0, MPI_COMM_WORLD);
 
   int rows_per_proc = N / size;
-  std::vector<int> sub_A(rows_per_proc * N);
+  std::vector<long long> sub_A(rows_per_proc * N);
 
-  MPI_Scatter(A.data(), rows_per_proc * N, MPI_INT, sub_A.data(), rows_per_proc * N, MPI_INT, 0, MPI_COMM_WORLD);
+  MPI_Scatter(A.data(), rows_per_proc * N, MPI_LONG_LONG, sub_A.data(), rows_per_proc * N, MPI_LONG_LONG, 0, MPI_COMM_WORLD);
 
-  std::vector<int> sub_C(rows_per_proc * N, 0);
+  std::vector<long long> sub_C(rows_per_proc * N, 0);
 
   double start_time = MPI_Wtime();
 
@@ -39,7 +39,7 @@ int main(int argc, char **argv)
       for (int k = 0; k < N; k++)
         sub_C[i * N + j] += sub_A[i * N + k] * B[k * N + j];
 
-  MPI_Gather(sub_C.data(), rows_per_proc * N, MPI_INT, C.data(), rows_per_proc * N, MPI_INT, 0, MPI_COMM_WORLD);
+  MPI_Gather(sub_C.data(), rows_per_proc * N, MPI_LONG_LONG, C.data(), rows_per_proc * N, MPI_LONG_LONG, 0, MPI_COMM_WORLD);
 
   double end_time = MPI_Wtime();
